cmp_resource: Rename models to model
