+++ components/org.wso2.carbon.identity.organization.resource.sharing.policy.management/src/main/java/org/wso2/carbon/identity/organization/resource/sharing/policy/management/util/ResourceSharingUtils.java
@@ -55,4 +55,6 @@ public class ResourceSharingUtils {
 
         return new ResourceSharingPolicyMgtServerException(error.getMessage(), error.getCode());
     }
+
+    private ResourceSharingUtils() {}
 }
